Hotfix/list of biographies (#25)
In this pull request, we have addressed the filtering issue within the list of biographies. We've updated the tagging for 'Convoy stories,' now marking them with the "convoy" tag. Previously, there was a problem where newly added Convoy stories were mistakenly mixed with Biography stories in the list, as the filtering mechanism excluded only those stories tagged as 'static.'